infra/pyproject.toml
[project]
name = "devsecops-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = [
    "A002",  # Allow 'id' argument (CDK pattern)
    "S106",  # Allow secret_name and token URLs (not actual secrets)
]
